+++ lib/enumerize/base.rb
@@ -91,11 +91,9 @@ module Enumerize
 
     def _set_default_value_for_enumerized_attributes
       self.class.enumerized_attributes.each do |attr|
-        # remove after dropping support for Rails 3.x
-        # https://github.com/brainspec/enumerize/issues/101
-        begin
+        if respond_to?(attr.name)
           attr_value = public_send(attr.name)
-        rescue ActiveModel::MissingAttributeError
+        else
           next
         end